+++ backend-nest/src/app.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { HealthController } from './health.controller';
 import { DatabaseModule } from './database/database.module';
 import { MarkerModule } from './marker/marker.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
-  imports: [DatabaseModule, MarkerModule],
+  imports: [DatabaseModule, MarkerModule, AuthModule],
   controllers: [HealthController]
 })
 export class AppModule {}